+++ contract-tests/tests/test/amazon/gen_ai/langchain/langchain_test.py
@@ -3,10 +3,14 @@
 from typing_extensions import override
 
 from amazon.gen_ai.gen_ai_test_base import (
+    AGENT_FINAL_OUTPUT,
+    GEN_AI_AGENT_NAME,
+    GEN_AI_OPERATION_NAME,
     GEN_AI_REQUEST_TEMPERATURE,
     GEN_AI_RESPONSE_ID,
     GEN_AI_RESPONSE_MODEL,
     GEN_AI_TOOL_DESCRIPTION,
+    GenAiOperationNameValues,
     GenAiProviderNameValues,
     GenAITestBase,
 )
@@ -35,10 +39,12 @@ class LangChainTest(GenAITestBase):
 
     @override
     def _assert_invoke_agent_spans(self, invoke_agent_spans: list, expected_count: int = 1):
-        super()._assert_invoke_agent_spans(invoke_agent_spans, expected_count)
+        self.assertEqual(len(invoke_agent_spans), expected_count)
         for span in invoke_agent_spans:
             attrs = self._get_attributes_dict(span.attributes)
-            self.assertIn(GEN_AI_REQUEST_TEMPERATURE, attrs)
+            self._assert_str_attribute(attrs, GEN_AI_OPERATION_NAME, GenAiOperationNameValues.INVOKE_AGENT.value)
+            self.assertIn(GEN_AI_AGENT_NAME, attrs)
+            self._assert_invoke_agent_content(attrs, span.name, AGENT_FINAL_OUTPUT)
 
     @override
     def _assert_execute_tool_spans(self, execute_tool_spans: list, expected_count: int = 1):